refactor(chart): restructure file to make it cleaner
