+++ packages/elements/src/chart/plugins/index.ts
@@ -1,0 +1,4 @@
+import doughnutCenterLabelPlugin from './doughnut-center-label.js';
+export {
+  doughnutCenterLabelPlugin
+};
